Use authentication helper to log in test users
