+++ app/controllers/logged_area_controller.rb
@@ -10,8 +10,6 @@ class LoggedAreaController < ApplicationController
   end
 
   def authenticate
-    return if Rails.env.test?
-
     authenticate_or_request_with_http_basic('Please sign in with username and password provided to you') do |username, password|
       if admin_login?(username, password)
         return true
