+++ spec/features/dataset_show_page_spec.rb
@@ -1,8 +1,9 @@
 require 'rails_helper'
 
 feature 'Dataset page', elasticsearch: true do
-
   before(:each) do
+    authenticate_browser
+
     stub_request(:any, FETCH_PREVIEW_URL).
       to_return(status: 200)
   end
