+++ spec/features/search_page_spec.rb
@@ -1,8 +1,11 @@
 # coding: utf-8
 require 'rails_helper'
 
-
 feature 'Search page', elasticsearch: true do
+  before(:each) do
+    authenticate_browser
+  end
+
   scenario 'Displays a not found message when a search returns 0 results' do
     query = 'interesting dataset'
